+++ packages/plugins/Draw/src/types.ts
@@ -4,7 +4,12 @@ import { Feature } from 'ol'
 import VectorLayer from 'ol/layer/Vector'
 import VectorSource from 'ol/source/Vector'
 import { StyleLike } from 'ol/style/Style'
-import { DrawConfiguration, DrawMode } from '@polar/lib-custom-types'
+import {
+  DrawConfiguration,
+  DrawMode,
+  MeasureMode,
+  MeasureOptions,
+} from '@polar/lib-custom-types'
 
 // The options that can be given to an ol/VectorLayer. Somehow the direct import from ol doesn't work.
 // This is a copy with the things that we currently use
@@ -29,19 +34,23 @@ export interface DrawState {
   selectedSize: number
   featureCollection: FeatureCollection
   selectedFeature: number
+  measureMode: MeasureMode
 }
 
 export interface DrawGetters extends Omit<DrawState, 'selectedFeature'> {
-  selectableDrawModes: DrawMode[]
+  selectableDrawModes: { [k in DrawMode]?: string }
   selectedFeature: Feature
-  selectableModes: Mode[]
+  selectableModes: { [k in Mode]: string }
+  selectableMeasureModes: { [k in MeasureMode]?: string }
   showTextInput: boolean
   configuration: DrawConfiguration
   fontSizes: number[]
+  measureOptions: MeasureOptions
   showSizeSlider: boolean
   /* actual text size to use */
   textSize: number
   showDrawOptions: boolean
+  showMeasureOptions: boolean
 }
 
 export interface DrawMutations extends MutationTree<DrawState> {