+++ packages/plugins/Gfi/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: Add new configuration parameter `multiSelect` to enable the possibility to choose between the selecting multiple features through a box or through a circle. The addition of this parameter deprecates the previously used parameter `boxSelect`.
 - Fix: Correctly disable `directSelect` if the user is currently using functionality of `@polar/plugin-draw`.
 
 ## 2.1.0